feat: add phpunit config, autoload-dev, and fix test structure   - Add phpunit.xml.dist with test suite pointing to test/ directory   - Add autoload-dev to composer.json mapping Test\\ namespace to test/   - Update composer test script to use phpunit directly (no hardcoded path)   - Extract FakeController into its own file (test/unit/FakeController.php)     to comply with PSR-2 one-class-per-file rule   - Fix CRLF line endings across src/ and test/ files
