+++ .gitignore
@@ -1,4 +1,5 @@
 .php-cs-fixer.cache
+.phpunit.result.cache
 vendor
 psalm.xml
 .idea
