+++ composer.json
@@ -8,6 +8,10 @@
     {
       "name": "jose garcia",
       "email": "jose.garcia051@gmail.com"
+    },
+    {
+      "name": "alejandro foucault",
+      "email": "ale@ajfm88.com"
     }
   ],
   "require": {
@@ -22,13 +26,18 @@
       "src/Garcia/helpers.php"
     ]
   },
+  "autoload-dev": {
+    "psr-4": {
+      "Test\\": "test/"
+    }
+  },
   "require-dev": {
     "phpunit/phpunit": "^9.0",
     "friendsofphp/php-cs-fixer": "^3.65",
     "squizlabs/php_codesniffer": "^3.7"
   },
     "scripts": {
-        "test": "phpunit --bootstrap vendor/autoload.php test/unit/RouterTest.php",
+        "test": "phpunit",
         "fix": "vendor/bin/php-cs-fixer fix ./ --rules=@PSR2",
         "lint": "vendor/bin/phpcs"
     }
